==> 1_Dataset_cleaning.ipynb <==
import pandas as pd
import string

df_openwebtext10 = pd.read_csv('datasets/openwebtext_10k_sent.csv')
df_openwebtext10

def special_symbols(sentence):
  symbols = '#$%&\*+/<=>'
  count = lambda symbols, set_symbols: sum([1 for x in symbols if x in set_symbols])
  return count(sentence, set(symbols))

def length_symbols(sentence):
  return len(sentence)

def length_words(sentence):
  return len(sentence.split())

def is_sentence(sentence):
  if type(sentence) != str: return 'Not sentence'
  count = lambda symbols, set_symbols: sum([1 for x in symbols if x in set_symbols])
  if count(sentence, set(string.ascii_letters)) < count(sentence, set(string.digits)): return 'Not sentence'
  if count(sentence, set(string.ascii_letters)) < count(sentence, set(string.punctuation)): return 'Not sentence'
  if count(sentence, set(string.ascii_letters)) < count(sentence, set(string.digits)) + count(sentence, set(string.punctuation)): return 'Not sentence'
  if (length_words(sentence) > 80) or (length_words(sentence) <= 3): return 'Not sentence'
  if length_symbols(sentence) < 11: return 'Not sentence'
  if ('@' in sentence) or ('http' in sentence): return 'Not sentence'
  if special_symbols(sentence) > 3: return 'Not sentence'
  return 'Sentence'

def is_complete(sentence):
  if sentence[-1] in '''!'".?»”''': return 'Complete'
  if sentence[-3] in '''!'".?»”''': return 'Complete by third'
  return 'Not complete'

def minus_two_symbols(sentence):
  if (sentence[-1] not in '''!'".?»”''') and (sentence[-3] in '''!'".?»”'''): return sentence[:-2]
  return sentence

df_openwebtext10['is_sentence'] = df_openwebtext10['text'].apply(is_sentence)
df_openwebtext10 = df_openwebtext10[df_openwebtext10['is_sentence'] != 'Not sentence']
df_openwebtext10['Complete'] = df_openwebtext10['text'].apply(is_complete)
df_openwebtext10 = df_openwebtext10[df_openwebtext10['Complete'] != 'Not complete']
df_openwebtext10['text'] = df_openwebtext10['text'].apply(minus_two_symbols)
df_openwebtext10 = df_openwebtext10.drop(columns=['is_sentence', 'Complete'])
df_openwebtext10 = df_openwebtext10.reset_index(drop=True)
df_openwebtext10

df_openwebtext10.to_csv('datasets/openwebtext_10k_sent_clean.csv', index=False)
