==> 3_Dataset_second_cleaning.ipynb <==
import pandas as pd

df_openwebtext10_with_attributes = pd.read_csv('datasets/openwebtext_10k_sent_with_attributes.csv')
df_openwebtext10_with_attributes

df_openwebtext10_with_attributes['pos'] = df_openwebtext10_with_attributes['Parts of speech'].apply(lambda x: ' '.join(x[2:-2].split("', '")))

expected_data_indexes = df_openwebtext10_with_attributes.apply(lambda row: (len(row['text'].split(' ')) == len(row['Parts of speech'].split(' ')) == row['Number of words']), axis=1)
print(sum(~expected_data_indexes))

df_openwebtext10_with_attributes = df_openwebtext10_with_attributes.drop(df_openwebtext10_with_attributes[~expected_data_indexes].index, axis=0)
df_openwebtext10_with_attributes = df_openwebtext10_with_attributes.drop(['pos'], axis=1)
df_openwebtext10_with_attributes

df_openwebtext10_with_attributes = df_openwebtext10_with_attributes.drop(df_openwebtext10_with_attributes[df_openwebtext10_with_attributes['Sentiment analysis'] == 'Error'].index)
df_openwebtext10_with_attributes

df_openwebtext10_with_attributes.to_csv('datasets/openwebtext_10k_sent_with_attributes_clean.csv', index=False)
